Fix/statemachine (#35)
* スタイルガイドとCLAUDE.mdに#include禁止の規約を追加し、state-pattern.mdに複数ファイルに分割したシーン管理のパターンを追加

* ステートマシン関連のヘッダーファイルを削除

* 警告レベルをLevel4に変更し、未使用の引数に[[maybe_unused]]を追加

* 警告レベルをLevel4に変更し、未使用の引数に[[maybe_unused]]を追加。進捗計算の精度を向上させるために、100を100.0に変更。

+++ HspppStateSample/StateSampleMain.cpp
@@ -108,11 +108,11 @@ void hspMain() {
           mes("Loading...");
           
           // ローディングバー風の演出
-          double progress = (sm.state_frame_count() * 100) / 120;
+          double progress = (sm.state_frame_count() * 100.0) / 120;
           color(100, 100, 100);
           boxf(170, 280, 470, 300);
           color(100, 200, 255);
-          boxf(170, 280, 170 + progress * 3, 300);
+          boxf(170, 280, 170 + static_cast<int>(progress * 3), 300);
           
           vwait();  // アニメーション画面: 60FPS
       });
